Converted to single-header, cleaned up and documented api a bit.
